# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "E-commerce backend. Money is always in minor units: a price_cents of 1999 means $19.99.",
        "title": "go-ecommerce API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/categories": {
            "get": {
                "description": "Public. Returns the active categories, ordered by name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "List categories",
                "responses": {
                    "200": {
                        "description": "Categories",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CategoryListEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin only. The category is created active.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Create a category",
                "parameters": [
                    {
                        "description": "Category details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CategoryEnvelope"
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid access token",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Not an administrator",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin only. Omitting is_active leaves it unchanged.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Update a category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.UpdateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated category",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CategoryEnvelope"
                        }
                    },
                    "400": {
                        "description": "Validation failed, or the id is not a number",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid access token",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Not an administrator",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "No such category",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin only. A category that still has products is refused; deactivate it instead, or move the products first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Delete a category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.MessageEnvelope"
                        }
                    },
                    "400": {
                        "description": "The id is not a number",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid access token",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Not an administrator",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "No such category",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "The category still has products",
                        "schema": {
                            "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CategoryEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CategoryResponse"
                },
                "message": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CategoryListEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CategoryResponse"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.CreateCategoryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "email is not available"
                },
                "message": {
                    "type": "string",
                    "example": "invalid credentials"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.MessageEnvelope": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "logged out"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "github_com_Hitesh-s0lanki_go-ecommerce_internal_dto.UpdateCategoryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "is_active": {
                    "description": "A pointer so an omitted field is distinguishable from false: without\nit, every update would silently deactivate the category.",
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access token from /auth/login, sent as \"Bearer \u003ctoken\u003e\". A refresh token is rejected here.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}